add option to specify the base directory used to compute exclusions from include option
for computation of exclusions from the include option always the "real" active directory
has to be used. When we switch the active and staging directories e.g. to use commit
instead of begin to reuse an existing staging directory, we must use this option to specify
which one is the "real" active directory